Refactor: Rename rhn.i18n to rhn.stringutils
