+++ client/rhel/spacewalk-client-tools/src/up2date_client/rpmUtils.py
@@ -16,7 +16,7 @@
 
 import os
 import rpm
-from rhn.i18n import sstr
+from rhn.stringutils import sstr
 from up2date_client import transaction
 from up2date_client import up2dateLog
