+++ client/tools/mgr-cfg/config_management/rhncfg_diff.py
@@ -23,7 +23,7 @@ from config_common import handler_base, utils, cfg_exceptions
 from config_common.rhn_log import log_debug, die
 from config_common.file_utils import f_date, ostr_to_sym
 from uyuni.common.usix import next
-from rhn.i18n import bstr, sstr
+from rhn.stringutils import bstr, sstr
 
 
 class Handler(handler_base.HandlerBase):
