+++ client/tools/mgr-push/rhnpush_main.py
@@ -39,7 +39,7 @@ from optparse import Option, OptionParser
 from rhn.connections import idn_ascii_to_puny
 
 from rhn import rpclib
-from rhn.i18n import sstr
+from rhn.stringutils import sstr
 from uyuni.common.rhn_pkg import InvalidPackageError, package_from_filename
 from uyuni.common.usix import raise_with_tb
